Module Type_generic.Ident

type t

Runtime identifier for a generic computation. This is essentially a string whose purpose is to give reasonable error messages in case the dependency requirements for a generic are not met at runtime.

The field called required is needed in order to build a generic computation module. It is used to establish a set up that would explicitly list all the computation that are required by an other computation to work.

Generic computations are a way to build dynamically some operations on types. It is possible to build computation on top of each other. This ident type will be the key to talk about other computations at the point of setting up the dependencies.