Module S.Computation

exported to build a computation on top of a previous one

type 'a t = 'a t

include Variant_and_record_intf.S with type 'a t := 'a t

include sig ... end

module Tag_internal : sig ... end

module Tag : sig ... end

Witness of a tag, that is an item in a variant type, also called an "applied variant Constructor"

module Variant_internal : sig ... end

module Variant : sig ... end

module Field_internal : sig ... end

module Field : sig ... end

Witness of a field, that is an item in a record type. The first parameter is the record type, the second is the type of the field. Example:

module Record_internal : sig ... end

module Record : sig ... end

val int : int t

val int32 : int32 t

val int64 : int64 t

val nativeint : nativeint t

val char : char t

val float : float t

val string : string t

val bytes : bytes t

val bool : bool t

val unit : unit t

val option : 'a t -> 'a option t

val list : 'a t -> 'a list t

val array : 'a t -> 'a array t

val lazy_t : 'a t -> 'a lazy_t t

val ref_ : 'a t -> 'a Stdlib.ref t

val function_ : 'a t -> 'b t -> ('a -> 'b) t

val tuple2 : 'a t -> 'b t -> ('a * 'b) t

val tuple3 : 'a t -> 'b t -> 'c t -> ('a * 'b * 'c) t

val tuple4 : 'a t -> 'b t -> 'c t -> 'd t -> ('a * 'b * 'c * 'd) t

val tuple5 : 'a t -> 'b t -> 'c t -> 'd t -> 'e t -> ('a * 'b * 'c * 'd * 'e) t

val record : 'a Record.t -> 'a t

val variant : 'a Variant.t -> 'a t

module Named : Named with type 'a computation := 'a t