Module Typerep_lib

module Make_typename : sig ... end

module Named_intf : sig ... end

module Std : sig ... end

module Std_internal : sig ... end

type-safe runtime type introspection

module Type_abstract : sig ... end

Abstract types helpers.

module Type_equal : sig ... end

module Type_generic : sig ... end

A computation is the type of an operation that can be applied to various different kind of types. It is expressed as a type with one parameter:

module Type_generic_intf : sig ... end

module Typename : sig ... end

runtime representation of the name of type 'a. Useful for representing types with a nominal notion of equality

module Typerep_obj : sig ... end

module Typerepable : sig ... end

module Variant_and_record_intf : sig ... end

Place holder for common Variants and Fields interface